+++ NEWS.md
@@ -7,7 +7,7 @@ editor_options:
 photobiologySensors 0.5.2
 =========================
 
-* Data objects rebuilt with 'photobiology' 0.13.2.
+* Data objects rebuilt with 'photobiology' 0.14.0.
 * Add data for the 13-channel AS7343 digital spectral sensor from ams-OSRAM.
 * Add data for the 11-channel AS7341 digital spectral sensor from ams-OSRAM.
 * Add data for the 3-channel AS7331 digital ultraviolet sensor from ams-OSRAM.